Count number of rows in import test
Since we moved to map comparison we no longer check the number of rows.
We use map comparison since we are not sure what order rows have been inserted.
Being unsure of the order of row insertion is a side-effect of our current solution
to the requirement to report all non-unique tokens.